feat: add value for enum
in api we assume that enum has an int value, that is not a case in java.
The java enums are objects, which are set constants. Not narrowed to an
integer type.
To make them work within our ecosystem that includes other technologies,
where enums are integer based, I added a int field value and functions
to obtain the value for serialization, and create an enum based on the
integer.

This is part of adding an andorid-jni feature.