feat(stub): add stub feature
adds an empty implementation of generated interface.
It fully supports handling properties (with the property change singal
emission).
It generates empty methods, to be filled with bussiness logic.
It may serve as an example or base implementation.

It allows easy implementing android service side when provideded as a
backend, hence all the helper file, facory, and service starter which
implicitely start the service.

This is part of adding the android-jni feature.