feat(adroid): Add client with project files
adds a client adapter, that implements the ServiceConnection and the api
interface.
A client is generated per api interface and connects with its Service
counterpart, that may be provided by diferent applications.
It requires the desired package that exposes the service to be given for
binding.
The service must be already started - the intent does not start it.
The client allows registering an event listener to get info about all
the changes that service informs about.
In that way it allows adding a jni version of that listener, which will
be added later.

This is part of adding the android-jni feature.