feat(example) add client example
adds an almost ready to start application for test purposes.
IMPORTANT it generates the functionality only for first interface.
That is for usability purpose, some tabs could be introduced for each
interface, but this make example more complex.
It genereates buttons to request some propety changes:
IMPORTANT the logic how to change it needs to be implemented by user,
similar for executing methods, the user must provide proper parameters.
It prints the result of events received from service: property changes,
singals receivied, method results.

This is part of adding the android-jni feature.