feat(example): add test service example skeleton
service uses a stub as the backend.
It is generated only for the first api interface.
User needs to fill the logic of changing properties and emitting
singals.
All the buttons and text views are already correctly connected.
It provides a buttons to start or stop service, change properties, emit
singals. The methods are also working, if the client sends a request it
is served, but the logic needs to be implemented in the stub i.e the
_impl module.

This is part of adding the android-jni feature.