feat(api): add gradle project files
adds necessary gradle files to allow the project to be build with e.g.
android studio.
The gradle was chosen because of compatibility with unreal build system,
as the main purpose of all the changes in the series of commits is to
allow easy data change between unreal and android.
Most of the modules will be used with unreal to achieve that.

This is part of adding the android-jni feature.