feat(android): add skeleton for service adapter
an android service is created for each interface.
It allows setting a backend that fullfills the interface.
The intention is to be albe to provide java-jni impl e.g. by unreal (but
it can bu also pure java implementation).
The backend is created by factory, so it is in same thread as the
service.
Note that the services are created in applications thread, not in a separate
ones, so having many interfaces and thus many services does not cause
the many threads overhead.
The messenger part is not yet there. It will be added with following
tasks.
The service must be started implicitely with
context.startService(theService) function.

This is part of adding the android-jni feature.