feat: improve comparing complex types
use better comparing functions for structures and arrays.
Use implemented equals() for the strucutres.
Tests requires now filling the structures with some data other than
default, cause before setting new value and expected notification it is
checked for equality (not set in that case).
With new compaision that checks the content not addres of object it
needs to have some values.