feat: add example app
Adds an app that uses composite build and builds all dependant modules.
The modules are in the diretory of rootproject.
Each module must provide a version.
The app just makes instances of all intefaces implementations:
stub and if android feature enabled also the client and service
versions.
Adds a task to run all unit tests.