chore: Simplify enum handling in {struct,interface}parcelable
Inside {{Camel .{Struct,Interface}.Name}}Parcelable(Parcel in) and
writeToParcel(Parcel dest, int flags) we're checking whether a field is
of KindType "enum", IsPrimitive or else.

As it turns out, the branches for (is KindType "enum") and else are the
same. Simplify the code by removing the (is KindType "enum") branch in
all cases.

This change doesn't impact generated code. It only cleans up the
template.