fix(jni): fix unbind from client
When client unbinds correctly, there is no "confirmation" form service
side. It is assumed that if client want's to unbind it will always work.
Therefore internal cleaning needs to be done in this case.
Call to doCleanupForUnbinding was added for this case.
Now the client listners are informed that the client is not bound
anymore.
Also the cleanup for unbinding should not clean the client messenger,
this will not change through the client lifetime, and is created only
once, when the client is constructed. With this cleaned the client was
not able to use messenger in case of re-binding.